refactor: move `Error` to `mod error`
